Another fix for the issues with C++ and 64bit platforms<
git-svn-id: https://devel.codership.com/svn/galera/trunk@848 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ gcache/src/GCache_memops.cpp
@@ -116,7 +116,7 @@ namespace gcache
 
         // We can reliably allocate continuous buffer which is twice as small
         // as total cache area. So compare to half the space
-        if (size < (size_cache >> 1)) {
+        if (size < ((size_t)size_cache >> 1)) {
 
             gu::Lock lock(mtx);
             void*    ptr;
@@ -155,7 +155,7 @@ namespace gcache
 
         // We can reliably allocate continuous buffer which is twice as small
         // as total cache area. So compare to half the space
-        if (size >= (size_cache >> 1)) return 0;
+        if (size >= ((size_t)size_cache >> 1)) return 0;
 
         BufferHeader* bh = BH(ptr) - 1;
